+++ XRTK-Core/Packages/com.xrtk.core/Interfaces/Providers/Controllers/IMixedRealityControllerDataProvider.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+﻿// Copyright (c) XRTK. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using System.Collections.Generic;
@@ -8,7 +8,7 @@ namespace XRTK.Interfaces.Providers.Controllers
     /// <summary>
     /// Mixed Reality Toolkit device definition, used to instantiate and manage a specific device / SDK
     /// </summary>
-    public interface IMixedRealityControllerDataProvider : IMixedRealityDataProvider
+    public interface IMixedRealityControllerDataProvider : IMixedRealityInputDataProvider
     {
         /// <summary>
         /// Retrieve all controllers currently registered with this device at runtime (if direct access is required)
