+++ XRTK-Core/Packages/com.xrtk.core/Interfaces/Providers/Speech.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 82be8b6f802afb643990670a2dc56acf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
