+++ XRTK-Core/Packages/com.xrtk.core/Providers/Speech/WindowsDictationDataProvider.cs
@@ -13,7 +13,7 @@ using XRTK.Utilities.Async;
 using UnityEngine.Windows.Speech;
 #endif // UNITY_STANDALONE_WIN || UNITY_WSA || UNITY_EDITOR_WIN
 
-namespace XRTK.Providers.Controllers.Speech
+namespace XRTK.Providers.Speech
 {
     /// <summary>
     /// Dictation data provider for Windows 10 based platforms.
@@ -59,7 +59,7 @@ namespace XRTK.Providers.Controllers.Speech
         public override void Update()
         {
             base.Update();
-            
+
             if (!Application.isPlaying) { return; }
 
             if (!isTransitioning && IsListening && !Microphone.IsRecording(deviceName) && dictationRecognizer.Status == SpeechSystemStatus.Running)
